refactor: move strategies into own files
