+++ packages/@ourworldindata/grapher/src/color/BinningStrategies.ts
@@ -1,77 +1,26 @@
 import {
+    BinningStrategy,
+    MidpointMode,
+    ResolvedBinningStrategy,
+} from "@ourworldindata/types"
+import {
     firstOfNonEmptyArray,
     lastOfNonEmptyArray,
-    normaliseToSingleDigitNumber,
-    numberMagnitude,
-    RequiredBy,
-    roundSigFig,
 } from "@ourworldindata/utils"
-import { min, quantile } from "d3"
+import { quantile } from "d3"
 import { sortedUniq } from "lodash-es"
 import * as R from "remeda"
 import { match, P } from "ts-pattern"
+import {
+    isLogBinningStrategy,
+    runLogBinningStrategy,
+} from "./BinningStrategyLogarithmic.js"
+import {
+    isEqualSizeBinningStrategy,
+    runEqualSizeBinningStrategy,
+} from "./BinningStrategyEqualSizeBins.js"
 
-/**
- * Strategies:
- * - The log scales result in log-like steps, e.g. 1, 2, 5, 10, ...
- *   They are fully defined given a minValue, maxValue and step size, and
- *   then generate as many bins as needed to cover the range.
- *   `log-auto` chooses the step size automatically in order to get a decent number of bins.
- * - The equal size bins result in evenly spaced steps, e.g. 0, 1, 2, 3, ...
- *   They are defined by a minValue, maxValue, and a rough target number of bins.
- *   `equalSizeBins-few-bins` chooses a small number of bins, while
- *   `equalSizeBins-many-bins` chooses a large number of bins.
- *   They then generate nice round bin thresholds given the input data.
- * - `equalSizeBins-percent` is a special case, where for data that looks like percent
- *   from 0% to 100% we want to mostly use 0%, 10%, 20%, etc. bins.
- */
-
-const logBinningStrategies = [
-    "log-auto",
-    "log-1-2-5",
-    "log-1-3",
-    "log-10",
-] as const
-
-const equalSizeBinningStrategies = [
-    "equalSizeBins-normal",
-    "equalSizeBins-few-bins",
-    "equalSizeBins-many-bins",
-    "equalSizeBins-percent",
-] as const
-
-export const automaticBinningStrategies = [
-    "auto",
-    ...logBinningStrategies,
-    ...equalSizeBinningStrategies,
-] as const
-
-type LogBinningStrategy = (typeof logBinningStrategies)[number]
-type EqualSizeBinningStrategy = (typeof equalSizeBinningStrategies)[number]
-
-type ResolvedLogBinningStrategy = Exclude<LogBinningStrategy, "log-auto">
-
-type BinningStrategy = (typeof automaticBinningStrategies)[number]
-
-type ResolvedBinningStrategy = Exclude<BinningStrategy, "auto">
-
-/**
- * Sometimes, we do have a midpoint in our data. In many cases, a natural midpoint is zero
- * (e.g. for year-over-year change, net migration, temperature anomaly, etc.), but other
- * midpoints also make sense (e.g. for sex ratio).
- * If we have a midpoint, then we want to account for it when binning, and generate bins
- * that are centered or symmetric around the midpoint.
- */
-export const binningMidpointModes = [
-    undefined, // Automatic
-    "none", // No midpoint
-    "symmetric", // Symmetric bins around a midpoint, with negBins = -1 * posBins
-    "same-num-bins", // Symmetric bins around a midpoint, with negBins.length = posBins.length
-    "asymmetric", // Bins around a midpoint, with negBins.length not necessarily equal to posBins.length
-] as const
-export type MidpointMode = (typeof binningMidpointModes)[number]
-
-interface BinningStrategyConfig {
+export interface BinningStrategyConfig {
     strategy: BinningStrategy
     minValue?: number
     maxValue?: number
@@ -83,7 +32,7 @@ interface BinningStrategyConfig {
     createBinForMidpoint?: boolean
 }
 
-interface ResolvedBinningStrategyConfig {
+export interface ResolvedBinningStrategyConfig {
     strategy: ResolvedBinningStrategy
     minValue?: number
     maxValue?: number
@@ -104,7 +53,10 @@ const AUTO_EQUAL_BINS_MAX_MAGNITUDE_DIFF = 1.2
 /**
  * Calculates the log10 difference between two numbers, i.e. compute x such that lowerValue * 10^x = upperValue.
  */
-const calcMagnitudeDiff = (lowerValue: number, upperValue: number): number => {
+export const calcMagnitudeDiff = (
+    lowerValue: number,
+    upperValue: number
+): number => {
     if (lowerValue > upperValue)
         throw new Error("lowerValue must be less than upperValue")
 
@@ -116,14 +68,6 @@ const countValues = (sortedValues: number[], value: number): number => {
     const lastIndex = R.sortedLastIndex(sortedValues, value)
     return lastIndex - firstIndex
 }
-
-const isEqualSizeBinningStrategy = (
-    strategy: BinningStrategy | ResolvedBinningStrategy
-): strategy is EqualSizeBinningStrategy => strategy.startsWith("equalSizeBins-")
-
-const isLogBinningStrategy = (
-    strategy: BinningStrategy | ResolvedBinningStrategy
-): strategy is LogBinningStrategy => strategy.startsWith("log-")
 
 export const pruneUnusedBins = (
     bins: number[],
@@ -249,7 +193,7 @@ export const hasValidConfigForBinningStrategy = (
  * minValue and maxValue may either be explicitly given, or automatically computed from the data.
  * If auto-computed, they are based on some very much heuristic rules, based on quantiles etc.
  */
-const computeMinMaxForStrategy = (
+export const computeMinMaxForStrategy = (
     strategy: ResolvedBinningStrategy,
     sortedValues: number[],
     conf?: Partial<ResolvedBinningStrategyConfig>
@@ -402,30 +346,6 @@ const runBinningStrategyAroundMidpoint = (
     return bins
 }
 
-/**
- * The equal-size binning strategies all operate with a range of target bins depending on the strategy.
- * For example, [5, 9] means that there should be at least 5 bins and at most 9.
- * If there is a midpoint, the number of bins needs to be adjusted downward, because we are potentially
- * creating up to twice as many bins overall.
- */
-const getTargetBinCountForEqualSizeBinsStrategy = (
-    strategy: EqualSizeBinningStrategy,
-    { hasMidpoint }: { hasMidpoint?: boolean } = {}
-): readonly [number, number] => {
-    return match(strategy)
-        .with("equalSizeBins-few-bins", () =>
-            hasMidpoint ? ([1, 3] as const) : ([2, 5] as const)
-        )
-        .with("equalSizeBins-normal", () =>
-            hasMidpoint ? ([3, 6] as const) : ([5, 9] as const)
-        )
-        .with(
-            P.union("equalSizeBins-many-bins", "equalSizeBins-percent"),
-            () => (hasMidpoint ? ([4, 8] as const) : ([8, 12] as const))
-        )
-        .exhaustive()
-}
-
 const runResolvedBinningStrategy = (
     conf: ResolvedBinningStrategyConfig,
     { hasMidpoint }: { hasMidpoint: boolean }
@@ -437,14 +357,10 @@ const runResolvedBinningStrategy = (
     )
     return match(conf.strategy)
         .when(isEqualSizeBinningStrategy, () =>
-            equalSizeBins({
-                minValue,
-                maxValue,
-                targetBinCount: getTargetBinCountForEqualSizeBinsStrategy(
-                    conf.strategy as EqualSizeBinningStrategy,
-                    { hasMidpoint }
-                ),
-            })
+            runEqualSizeBinningStrategy(
+                { ...conf, minValue, maxValue },
+                { hasMidpoint }
+            )
         )
         .when(isLogBinningStrategy, () =>
             runLogBinningStrategy(
@@ -453,65 +369,6 @@ const runResolvedBinningStrategy = (
             )
         )
         .exhaustive()
-}
-
-/**
- * Automatically chooses a log binning strategy based on the magnitude difference between the min and max values.
- * Very roughly, the resulting number of bins is roughly magnitudeDiff * numberOfLogSteps.
- */
-const autoChooseLogBinningStrategy = (
-    magnitudeDiff: number
-): ResolvedLogBinningStrategy => {
-    if (magnitudeDiff >= 3.6) {
-        return "log-10"
-    }
-    if (magnitudeDiff >= 2.6) {
-        return "log-1-3"
-    }
-    return "log-1-2-5"
-}
-
-const runLogBinningStrategy = (
-    conf: RequiredBy<ResolvedBinningStrategyConfig, "minValue" | "maxValue">,
-    { hasMidpoint }: { hasMidpoint?: boolean } = {}
-): number[] => {
-    const { minValue, maxValue } = conf
-
-    if (minValue <= 0 || maxValue <= 0) {
-        throw new Error("Log binning strategy only supports positive values")
-    }
-
-    let resolvedStrategy: ResolvedLogBinningStrategy
-    if (conf.strategy === "log-auto") {
-        let magnitudeDiff = calcMagnitudeDiff(minValue, maxValue)
-        if (hasMidpoint) magnitudeDiff *= 1.8 // If there is a midpoint, we want to create fewer bins than we would otherwise
-        resolvedStrategy = autoChooseLogBinningStrategy(magnitudeDiff)
-    } else {
-        resolvedStrategy = conf.strategy as ResolvedLogBinningStrategy
-    }
-
-    const bins = match(resolvedStrategy)
-        .with("log-10", () =>
-            fakeLogBins({ minValue, maxValue, logSteps: [1] })
-        )
-        .with("log-1-3", () =>
-            fakeLogBins({ minValue, maxValue, logSteps: [1, 3] })
-        )
-        .with("log-1-2-5", () =>
-            fakeLogBins({ minValue, maxValue, logSteps: [1, 2, 5] })
-        )
-        .exhaustive()
-
-    // Add the midpoint (most likely zero) if it is not already included in a bin
-    if (
-        !hasMidpoint &&
-        conf.midpoint !== undefined &&
-        bins[0] > conf.midpoint
-    ) {
-        bins.unshift(conf.midpoint)
-    }
-
-    return bins
 }
 
 const autoChooseBinningStrategy = (
@@ -569,36 +426,6 @@ const autoChooseBinningStrategy = (
     return "log-auto"
 }
 
-const fakeLogBins = ({
-    minValue,
-    maxValue,
-    logSteps,
-}: {
-    minValue: number
-    maxValue: number
-    logSteps: number[]
-}): number[] => {
-    if (minValue <= 0 || maxValue <= 0) {
-        throw new Error("fakeLogBins only supports positive values")
-    }
-
-    const magnitudeMin = numberMagnitude(minValue) - 1
-    const magnitudeMax = numberMagnitude(maxValue) - 1
-
-    const candidates = R.range(magnitudeMin, magnitudeMax + 1).flatMap(
-        (magnitude) => {
-            const factor = Math.pow(10, magnitude)
-            return logSteps.map((step) => step * factor)
-        }
-    )
-
-    if ((R.last(candidates) ?? 0) < maxValue) {
-        candidates.push(1 * Math.pow(10, magnitudeMax + 1))
-    }
-
-    return pruneUnusedBins(candidates, { minValue, maxValue })
-}
-
 export const mirrorBinsAroundMidpoint = (
     binOffsets: number[],
     midpoint: number
@@ -607,51 +434,4 @@ export const mirrorBinsAroundMidpoint = (
     const leftBins = filteredRightBins.map((v) => midpoint - v).reverse()
     const rightBins = filteredRightBins.map((v) => v + midpoint)
     return [...leftBins, midpoint, ...rightBins]
-}
-
-export const equalSizeBins = ({
-    minValue,
-    maxValue,
-    targetBinCount,
-}: {
-    minValue: number
-    maxValue: number
-    targetBinCount: readonly number[]
-}): number[] => {
-    if (minValue > maxValue) {
-        throw new Error("minValue must be less than maxValue")
-    }
-
-    const range = maxValue - minValue
-
-    // Normalise the range to be within 1 and 10
-    const normalisedRange = normaliseToSingleDigitNumber(range)
-    // This is the factor used to un-normalise the values back to their original scale
-    const factor = Math.pow(10, numberMagnitude(range) - 1)
-
-    // These are all common and "good" step sizes; find the first one that gives us the target bin count
-    const stepSizeCandidates = [1, 0.1, 0.5, 0.2, 2, 3, 0.3, 0.75, 0.25]
-    const stepSize = stepSizeCandidates.find((candidateStepSize) => {
-        if (normalisedRange === 0) return true
-
-        const numSteps = Math.ceil(normalisedRange / candidateStepSize)
-        return numSteps >= targetBinCount[0] && numSteps <= targetBinCount[1]
-    })
-
-    if (stepSize === undefined) {
-        throw new Error("No valid step size found")
-    }
-
-    const steps = Math.ceil(normalisedRange / stepSize)
-
-    const actualStepSize = stepSize * factor
-
-    // Round min value to step size
-    const minValueRounded =
-        Math.floor(minValue / actualStepSize) * actualStepSize
-
-    return R.range(0, steps + 1).map((i) => {
-        const value = i * actualStepSize
-        return minValueRounded + roundSigFig(value, 3) // to avoid floating point issues
-    })
 }
