+++ packages/@ourworldindata/grapher/src/index.ts
@@ -115,3 +115,4 @@ export {
 } from "./core/GrapherUseHelpers.js"
 export { GeoFeatures } from "./mapCharts/GeoFeatures"
 export { isValidVerticalComparisonLineConfig } from "./comparisonLine/ComparisonLineHelpers"
+export { hasValidConfigForBinningStrategy } from "./color/BinningStrategies"
